@@ -1,9 +1,9 @@
 name: flutter_app
-description: A new Flutter application.
+description: "A new Flutter application."
 version: 4.6.5+465
 
 environment:
-  sdk: '>=2.12.0 <3.0.0'
+  sdk: ^3.3.0
 
 dependencies:
   flutter:
@@ -14,51 +14,52 @@ dependencies:
   #网络
   dio: ^5.4.0
   #  webview
-  flutter_inappwebview: ^6.0.0
+  flutter_inappwebview: ^6.1.5
   # 弃用（无法支持null-safety）
   # webview_cookie_manager: ^2.0.6
   #状态管理
   flutter_bloc: ^8.1.3
   #  图片
-  cached_network_image: ^3.0.0
+  cached_network_image: ^3.4.1
   #语言包
   flutter_localizations:
     sdk: flutter
   #包信息
-  package_info_plus: ^5.0.1
+  package_info_plus: ^8.1.2
   #url启动
 #  url_launcher: ^6.2.5
   #shared_preferences插件
-  shared_preferences: ^2.0.6
+  shared_preferences: ^2.3.4
   #  瀑布流
-  flutter_staggered_grid_view: ^0.4.0
+  flutter_staggered_grid_view: ^0.7.0
   #二维码
 #  qr_flutter: ^4.0.0
-  intl: ^0.18.0
+  intl: ^0.19.0
   #轮播
-  carousel_slider: ^4.0.0
+  carousel_slider: ^5.0.0
   #高性能json序列化
-  json_annotation: ^4.3.0
+  json_annotation: ^4.9.0
 
-#  flutter_html: ^2.2.1
-
+  #  flutter_html: ^2.2.1
 
   #  对话框
   bot_toast: ^4.1.3
-  event_bus: 2.0.0
+  event_bus: ^2.0.1
   flutter_datetime_picker: ^1.5.1
-#  qr_code_scanner: ^0.5.2
+  #  qr_code_scanner: ^0.5.2
   video_player: ^2.1.12
   permission_handler: ^11.1.0
-  path_provider: ^2.0.2
-#  install_plugin: ^2.0.1=
+#  path_provider: ^2.1.5
+  #  install_plugin: ^2.0.1=
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  build_runner: ^2.1.4
+
+  flutter_lints: ^5.0.0
   #  json序列化需要的库，只需要dev
-  json_serializable: ^6.0.0
+  build_runner: ^2.1.8
+  json_serializable: ^6.1.0
 
 flutter:
 
