@@ -13,13 +13,13 @@ Cells _$CellsFromJson(Map<String, dynamic> json) => Cells()
   ..title = json['title'] as String?
   ..picUrl = json['picUrl'] as String?
   ..rcmdSort = json['rcmdSort'] as bool?
-  ..itemCnt = json['itemCnt'] as int?
-  ..itemFrom = json['itemFrom'] as int?
+  ..itemCnt = (json['itemCnt'] as num?)?.toInt()
+  ..itemFrom = (json['itemFrom'] as num?)?.toInt()
   ..subTitle = json['subTitle'] as String?
   ..titleColor = json['titleColor'] as String?
   ..showPrice = json['showPrice'] as bool?
   ..popupUrl = json['popupUrl'] as String?
-  ..leftTime = json['leftTime'] as int?
+  ..leftTime = (json['leftTime'] as num?)?.toInt()
   ..id = json['id']
   ..targetUrl = json['targetUrl'] as String?
   ..itemList = (json['itemList'] as List<dynamic>?)
