@@ -15,18 +15,18 @@ CouponItemModel _$CouponItemModelFromJson(Map<String, dynamic> json) =>
       ..validStartTime = json['validStartTime'] as num?
       ..appOnly = json['appOnly'] as bool?
       ..buttonFlag = json['buttonFlag'] as bool?
-      ..type = json['type'] as int?
+      ..type = (json['type'] as num?)?.toInt()
       ..path = json['path'] as String?
       ..usable = json['usable'] as bool?
       ..unit = json['unit'] as String?
       ..hasCoupon = json['hasCoupon'] as bool?
-      ..newUserOnly = json['newUserOnly'] as int?
+      ..newUserOnly = (json['newUserOnly'] as num?)?.toInt()
       ..validEndTime = json['validEndTime'] as num?
       ..briefDesc = json['briefDesc'] as String?
       ..useCondition = json['useCondition'] as String?
       ..name = json['name'] as String?
       ..timeDesc = json['timeDesc'] as String?
-      ..useTime = json['useTime'] as int?
+      ..useTime = (json['useTime'] as num?)?.toInt()
       ..id = json['id'] as num?
       ..cash = (json['cash'] as num?)?.toDouble()
       ..albeToActivated = json['albeToActivated'] as bool?;
