@@ -22,7 +22,7 @@ CarItem _$CarItemFromJson(Map<String, dynamic> json) => CarItem()
   ..editChecked = json['editChecked'] as bool?
   ..promNotSatisfyType = json['promNotSatisfyType'] as num?
   ..promotionBtn = json['promotionBtn'] as num?
-  ..allowCount = json['allowCount'] as int?
+  ..allowCount = (json['allowCount'] as num?)?.toInt()
   ..type = json['type'] as num?
   ..suitCount = json['suitCount'] as num?
   ..source = json['source'] as num?
