@@ -38,7 +38,7 @@ CartItemListItem _$CartItemListItemFromJson(Map<String, dynamic> json) =>
       ..limitPurchaseCount = json['limitPurchaseCount'] as num?
       ..preemptionStatus = json['preemptionStatus'] as num?
       ..canSwitchSpec = json['canSwitchSpec'] as bool?
-      ..stepNo = json['stepNo'] as int?
+      ..stepNo = (json['stepNo'] as num?)?.toInt()
       ..checkExt = json['checkExt']
       ..specList = (json['specList'] as List<dynamic>?)
           ?.map((e) => SpecListItem.fromJson(e as Map<String, dynamic>))
